+++ emstrack/models.py
@@ -2,43 +2,12 @@ from django.contrib.auth.models import User
 
 from django.contrib.gis.db import models
 from django.contrib.gis.geos import Point
-from django.template.defaulttags import register
 from django.utils.translation import ugettext_lazy as _
 
-# filters
 from django.utils import timezone
-from django.utils.safestring import mark_safe
 
-from django.conf import settings
-
-
-@register.filter(is_safe=True)
-def get_check(key):
-    if key:
-        return mark_safe('<span class="fas fa-check"></span>')
-    else:
-        return ''
-
-
-@register.filter(is_safe=True)
-def get_times(key):
-    if key:
-        return ''
-    else:
-        return mark_safe('<span class="fas fa-times"></span>')
-
-
-@register.filter(is_safe=True)
-def get_check_or_times(key):
-    if key:
-        return mark_safe('<span class="fas fa-check"></span>')
-    else:
-        return mark_safe('<span class="fas fa-times"></span>')
-
-
-@register.simple_tag
-def is_video_enabled():
-    return settings.ENABLE_VIDEO
+# filters
+import emstrack.filters
 
 
 defaults = {
